PUTBASE OK OMG

+++ libftprintf/libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: sgaudin <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/01/25 14:18:19 by sgaudin           #+#    #+#             */
-/*   Updated: 2016/01/25 14:19:22 by sgaudin          ###   ########.fr       */
+/*   Updated: 2016/01/25 15:17:56 by sgaudin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,6 @@
 
 # include "libft.h"
 
-int			ftp_putbase(uint8_t n, uint32_t base);
+int			ftp_putbase(uint32_t n, uint32_t base);
 
 #endif
